Re-organize Dockerfile layers
Copy Composer executable and source code after installing the OS
dependencies to keep that layer cached on code changes. Ideally,
we'd copy the `composer.json` and lock files first to do the same
with the Composer dependencies but they're currently installed
inside the project and therefore overwritten when the source code
is copied.